+++ src/camera.h
@@ -6,14 +6,14 @@ class Camera
 {
 public:
   Camera();
-  Camera(int pixel_nx, int pixel_ny);
+  Camera(float pixel_nx, float pixel_ny);
 
   // camera needs to know pixel image dimensions...
   // so, you can provide to the Camera constructor or use a default,
   // or even pass in as an argument to the generateRay function
   // virtual ray generateRay( int i, int j, int nx, int ny ) = 0;
 
-  virtual ray generateRay(int i, int j) = 0;
+  virtual ray generateRay(float i, float j) = 0;
 
 protected:
   vec3 pos;// position of camera
@@ -29,5 +29,5 @@ protected:
   float focalLength;// also sometimes referred to as "d"
   float imagePlane_width, imagePlane_height;
 
-  int nx, ny;
+  float nx, ny;
 };
